move fetch into standalone package
