+++ mars/tensor/execution/fetch.py
@@ -1,6 +1,5 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# Copyright 1999-2017 Alibaba Group Holding Ltd.
+# Copyright 1999-2018 Alibaba Group Holding Ltd.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,17 +13,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .core import Operand
-from .. import opcodes as OperandDef
-from ..serialize import StringField
-from ..utils import to_str
+from ..expressions import fetch
 
 
-class Fetch(Operand):
-    _op_type_ = OperandDef.FETCH
+def register_fetch_handler():
+    from ...executor import register
 
-    _to_fetch_key = StringField('to_fetch_key', on_serialize=to_str)
-
-    @property
-    def to_fetch_key(self):
-        return self._to_fetch_key
+    register(fetch.TensorFetch, lambda *_: None)
