+++ mars/worker/execution.py
@@ -25,7 +25,7 @@ from ..compat import six, Enum, BrokenPipeError, ConnectionRefusedError, Timeout
 from ..config import options
 from ..errors import PinChunkFailed, WorkerProcessStopped, ExecutionInterrupted, \
     DependencyMissing
-from ..tensor.expressions.datasource import TensorFetch
+from ..operands import Fetch
 from ..utils import deserialize_graph, log_unhandled, to_str
 from .chunkholder import ensure_chunk
 from .spill import spill_exists, get_spill_data_size
@@ -256,7 +256,7 @@ class ExecutionActor(WorkerActor):
                 continue
 
             missing_keys = [c.key for c in succ_rec.graph if c.key not in succ_rec.data_sizes
-                            and isinstance(c.op, TensorFetch)]
+                            and isinstance(c.op, Fetch)]
             if missing_keys:
                 sizes = self.get_meta_ref(session_id, graph_key, local=False) \
                     .batch_get_chunk_size(session_id, missing_keys)
@@ -294,7 +294,7 @@ class ExecutionActor(WorkerActor):
 
         # collect potential allocation sizes
         for chunk in graph:
-            if not isinstance(chunk.op, TensorFetch) and chunk.key in graph_record.targets:
+            if not isinstance(chunk.op, Fetch) and chunk.key in graph_record.targets:
                 # use estimated size as potential allocation size
                 alloc_mem_batch[chunk.key] = chunk.rough_nbytes * 2
                 alloc_cache_batch[chunk.key] = chunk.rough_nbytes
@@ -432,7 +432,7 @@ class ExecutionActor(WorkerActor):
         graph_record = self._graph_records[session_graph_key]
         graph = graph_record.graph
 
-        ops = set(type(c.op).__name__ for c in graph if not isinstance(c.op, TensorFetch))
+        ops = set(type(c.op).__name__ for c in graph if not isinstance(c.op, Fetch))
         op_calc_key = ('calc_speed.' + list(ops)[0]) if len(ops) == 1 else None
 
         stats = defaultdict(lambda: dict(count=0))
@@ -454,7 +454,7 @@ class ExecutionActor(WorkerActor):
 
         if calc_fetch:
             for c in graph:
-                if not isinstance(c.op, TensorFetch):
+                if not isinstance(c.op, Fetch):
                     break
                 input_size += c.nbytes
                 if self._chunk_holder_ref.is_stored(c.key):
@@ -587,7 +587,7 @@ class ExecutionActor(WorkerActor):
 
         handled_keys = set()
         for chunk in graph_record.graph:
-            if not isinstance(chunk.op, TensorFetch):
+            if not isinstance(chunk.op, Fetch):
                 continue
             if chunk.key in handled_keys:
                 continue
@@ -659,7 +659,7 @@ class ExecutionActor(WorkerActor):
             # get allocation for calc, in case that memory exhausts
             target_allocs = dict()
             for chunk in graph_record.graph:
-                if isinstance(chunk.op, TensorFetch):
+                if isinstance(chunk.op, Fetch):
                     if not self._chunk_holder_ref.is_stored(chunk.key):
                         alloc_key = self._build_load_key(graph_key, chunk.key)
                         if alloc_key in graph_record.mem_request:
